requirements.txt
easyocr
torch
torchvision
torchaudio
pyautogui
playwright
pillow